saves the volume in the credentials, closes #32

+++ src/main/java/com/shuffle/scplayer/core/AudioListener.java
@@ -32,7 +32,7 @@ public interface AudioListener {
 
     int onAudioData(byte[] data);
 
-    void onVolumeChanged(short volume);
+    void onVolumeChanged(int volume);
 
     void close();
 }
