+++ src/main/java/com/shuffle/scplayer/core/SpotifyConnectPlayerImpl.java
@@ -288,7 +288,7 @@ public class SpotifyConnectPlayerImpl implements SpotifyConnectPlayer {
 			}
 		};
 		spPlaybackCallbacks.apply_volume = new SpPlaybackCallbacks.apply_volume_callback() {
-			public void apply(short volume, Pointer userdata) {
+			public void apply(int volume, Pointer userdata) {
 				for (PlayerListener playerListener : playerListeners) {
 					playerListener.onVolumeChanged(volume);
 				}
@@ -383,7 +383,7 @@ public class SpotifyConnectPlayerImpl implements SpotifyConnectPlayer {
 	}
 
 	@Override
-	public short getVolume() {
+	public int getVolume() {
 		try {
 			libLock.lock();
 			return spotifyLib.SpPlaybackGetVolume();
@@ -519,7 +519,7 @@ public class SpotifyConnectPlayerImpl implements SpotifyConnectPlayer {
 	}
 
 	@Override
-	public void volume(short volume) {
+	public void volume(int volume) {
 		try {
 			libLock.lock();
 			spotifyLib.SpPlaybackUpdateVolume(volume);
